Langgraph/travel_planner/requirements.txt
pydantic 
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain_huggingface
faiss-cpu
langchain_google_genai
langchain-pinecone
langchain
weaviate-client
langchain-weaviate
youtube-search
youtube-search-python
langgraph
chromadb
langchain-chroma
duckduckgo-search
graphviz
yfinance
python-docx
groq
streamlit
openai
google-generativeai